+++ src/main/java/gift/model/SiteUser.java
@@ -1,4 +1,4 @@
-package gift.user;
+package gift.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
