+++ src/main/java/gift/user/UserSecurityService.java
@@ -1,5 +1,6 @@
 package gift.user;
 
+import gift.model.SiteUser;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
