+++ src/main/java/gift/user/UserService.java
@@ -1,5 +1,6 @@
 package gift.user;
 
+import gift.model.SiteUser;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
